+++ docs/changelog.rst
@@ -24,8 +24,10 @@ Changelog
   * Project
   * Issue
 
+- Added: Integer representation to all resources, i.e. __int__()
 - Added: Informal string representation to all resources, i.e. __str__()
 - Fixed: Conversion of issue priorities to enumeration resource object didn't work
+- Changed: Renamed version attribute to redmine_version in all resources to avoid name intersections
 - Changed: ResourceManager get() method now throws a ``ResourceNotFoundError`` exception if
   resource wasn't found instead of returning None in previous versions
 - Changed: reimplemented fix for __repr__() from 0.1.1